EmptyAppendTransaction - move upper level
